added fist and tied collision detection to it
