+++ Assets/FistMovement.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 755419156e2d948f6a769883bedbe075
+guid: 63d4fc78c1c684bf0873411020cb5a71
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
