+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,9 +6,9 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/MenuScreen.unity
+    path: Assets/Scenes/MenuScreen.unity
     guid: 2cda990e2423bbf4892e6590ba056729
   - enabled: 1
-    path: Assets/Level_1.unity
+    path: Assets/Scenes/Level_1.unity
     guid: ae25dc308e1334477845224ab2c644f3
   m_configObjects: {}
